backend/pyproject.toml
[tool.poetry]
name = "eureka-backend"
version = "0.1.0"
description = "Eureka MVP backend"
authors = ["Eureka Team <team@eureka.local>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = { extras = ["email"], version = "^2.9.2" }
pydantic-settings = "^2.5.2"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
bcrypt = "4.0.1"
python-multipart = "^0.0.12"
boto3 = "^1.35.0"
openpyxl = "3.1.2"
pandas = "2.2.2"
pdfplumber = "0.11.4"
xlrd = "2.0.1"
python-magic = "0.4.27"
chardet = "5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
reportlab = "^4.2.2"
xlwt = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
